save Appointment

+++ src/main/java/com/teethcare/model/entity/Booking.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -41,7 +42,7 @@ public class Booking {
     private Patient patient;
 
     @Column(name = "total_price")
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 
     @Column(name = "create_booking_date")
     private Timestamp createBookingDate;
